moved TransactionType to consts.py, created consts.PaymentPlanType
