+++ braspag/core.py
@@ -11,12 +11,9 @@ import jinja2
 
 from .utils import spaceless
 from .utils import is_valid_guid
-from .utils import method_must_be_redesigned
 from .exceptions import BraspagException
 from .exceptions import HTTPTimeoutError
 from .response import CreditCardAuthorizationResponse
-from .response import BilletResponse
-from .response import BilletDataResponse
 from .response import CreditCardCancelResponse
 from .response import CreditCardRefundResponse
 from .response import BraspagOrderIdResponse
@@ -27,6 +24,8 @@ from .response import BraspagOrderDataResponse
 from .response import AddCardResponse
 from .response import InvalidateCardResponse
 from .response import BraspagOrderIdDataResponse
+from .consts import TransactionType
+from .consts import PaymentPlanType
 from xml.dom import minidom
 
 from tornado.httpclient import HTTPRequest
@@ -35,16 +34,7 @@ from tornado import httpclient
 from tornado import gen
 
 
-class TransactionType(object):
-    PRE_AUTHORIZATION = '1'
-    AUTOMATIC_CAPTURE = '2'
-    PRE_AUTHORIZATION_WITH_AUTHENTICATION = '3'
-    AUTOMATIC_CAPTURE_WITH_AUTHENTICATION = '4'
-    RECURRENT_PRE_AUTHORIZATION = '5'
-    RECURRENT_AUTOMATIC_CAPTURE = '6'
-
-
-class GenericRequest(object):
+class BaseRequest(object):
     def __init__(self, merchant_id=None, homologation=False, request_timeout=10):
         self.merchant_id = merchant_id
 
@@ -105,7 +95,7 @@ class GenericRequest(object):
         try:
             body = minidom.parseString(payload.encode('utf-8')).toprettyxml(indent='  ')
         except Exception as e:
-            body = response.body
+            body = payload
         return body
 
     @gen.coroutine
@@ -115,11 +105,16 @@ class GenericRequest(object):
             response = yield self.http_client.fetch(self._get_request(url, xml))
         except HTTPError as e:
             raise e.code == 599 and HTTPTimeoutError(e.code, e.message) or HTTPError(e.code, e.message)
-        self.log.debug('Response code: %s body: %s' % (response.code, self.pretty_body(response.body)))
+
+        if not response:
+            self.log.debug('Response code: %s body: %s' % (response.code, self.pretty_body(response.body)))
+        else:
+            self.log.error('No response received.')
+
         raise gen.Return(response)
 
 
-class BraspagRequest(GenericRequest):
+class BraspagRequest(BaseRequest):
     """
     Implements Braspag Pagador API (manual version 1.9).
     """
@@ -320,71 +315,6 @@ class BraspagRequest(GenericRequest):
                                                     context), query=True)
         raise gen.Return(BraspagOrderIdDataResponse(response.body))
 
-    @method_must_be_redesigned
-    def issue_billet(self, **kwargs):  # pragma: no cover
-        """DEPRECATED -- must be redesigned to work asynchronously.
-
-        All arguments supplied to this method must be keyword arguments.
-
-        :arg order_id: Order id. It will be used to indentify the
-                       order later in Braspag.
-        :arg customer_id: Must be user's CPF/CNPJ.
-        :arg customer_name: User's full name.
-        :arg customer_email: User's email address.
-        :arg amount: Amount to charge.
-        :arg currency: Currency of the given amount. *Default: BRL*.
-        :arg country: User's country. *Default: BRA*.
-        :arg payment_method: Payment method code
-        :arg soft_descriptor: Order description to be shown on the customer
-                              billet. Maximum of 13 characters.
-
-        :returns: :class:`~braspag.BraspagResponse`
-
-        """
-        if not kwargs.get('currency'):
-            kwargs['currency'] = 'BRL'
-
-        if not kwargs.get('country'):
-            kwargs['country'] = 'BRA'
-
-        soft_desc = ''
-        if kwargs.get('soft_descriptor'):
-            # only keep first 13 chars
-            soft_desc = kwargs.get('soft_descriptor')[:13]
-
-            # Replace special chars by ascii
-            soft_desc = unicodedata.normalize('NFKD', soft_desc)
-            soft_desc = soft_desc.encode('ascii', 'ignore')
-
-        kwargs['soft_descriptor'] = soft_desc
-
-        kwargs['is_billet'] = True
-
-        xml_request = self._render_template('authorize_billet.xml', kwargs)
-        return BilletResponse(self._request(spaceless(xml_request)))
-
-    @method_must_be_redesigned
-    def get_billet_data(self, **kwargs):  # pragma: no cover
-        """DEPRECATED -- must be redesigned to work asynchronously.
-
-        All arguments supplied to this method must be keyword arguments.
-
-        :arg transaction_id: The id of the transaction generated previously by
-        *issue_billet*
-
-        :returns: :class:`~braspag.BilletResponse`
-
-        """
-        assert is_valid_guid(kwargs.get('transaction_id')), 'Invalid Transaction ID'
-
-        context = {
-            'transaction_id': kwargs.get('transaction_id'),
-            'request_id': kwargs.get('request_id')
-        }
-        xml_request = self._render_template('get_billet_data.xml', context)
-        xml_response = self._request(spaceless(xml_request), query=True)
-        return BilletDataResponse(xml_response)
-
 
 class BraspagTransaction(object):
     """
@@ -439,13 +369,10 @@ class BraspagTransaction(object):
             raise BraspagException('Number of payments must be int.')
 
         if not kwargs.get('payment_plan'):
-
             if number_of_payments > 1:
-                # 2 = parcelado pelo emissor do cartão
-                kwargs['payment_plan'] = 2
+                kwargs['payment_plan'] = PaymentPlanType.INSTALLMENTS_BY_ESTABLISHMENT
             else:
-                # 0 = a vista
-                kwargs['payment_plan'] = 0
+                kwargs['payment_plan'] = PaymentPlanType.NO_INSTALLMENTS
 
         if not kwargs.get('currency'):
             kwargs['currency'] = 'BRL'
@@ -480,7 +407,7 @@ class BraspagTransaction(object):
             setattr(self, attr, kwargs[attr])
 
 
-class ProtectedCardRequest(GenericRequest):
+class ProtectedCardRequest(BaseRequest):
     """
     Implements Braspag Cartão Protegido API (manual version 2.1).
     """
